optimize interpolation crate
complete two todos, move barycentric_weight and interpolate to interpolation crate
